+++ dto/debt-dto.go
@@ -1,7 +1,7 @@
 package dto
 
 type DebtDTO struct {
-	Amount      int    `json:"amount" binding:"required,gt=0"`
-	Description string `json:"description" binding:"required"`
-	DebtorID    string `json:"debtorId" binding:"omitempty"`
+	Amount      float32 `json:"amount" binding:"required,gt=0"`
+	Description string  `json:"description" binding:"required"`
+	DebtorID    string  `json:"debtorId" binding:"omitempty"`
 }
